Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

ARCTR.TrustRegionException

ARCTR.compute_r

ARCTR.ldlt_symm

​

ARCTR.TrustRegionException — Type

Exception type raised in case of error.

ARCTR.compute_r — Method

Compute the actual vs. predicted reduction radio ∆f/Δm, where Δf = f - ftrial is the actual reduction is an objective/merit/penalty function, Δq = q - qtrial is the reduction predicted by the model q of f. We assume that q is being minimized, and therefore that Δq > 0.

ARCTR.ldlt_symm — Method

Higams' ldlt_symm translated from matlab. Performs a so called BKK Bounded Bunch Kaufman factorization of A0, that means ||L|| is bounded and bounded away from zero.