Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

ARCTR.TrustRegionException

ARCTR.compute_r

​

ARCTR.TrustRegionException — Type

Exception type raised in case of error.

ARCTR.compute_r — Method

r, good_grad, gnext = compute_r(nlp, f, Δf, Δq, slope, d, xnext, gnext, robust)

Arguments:

nlp: Current model we are trying to solve

f: current objective value

Δf: = f - f_trial is the actual reduction is an objective/merit/penalty function,

Δq: q - q_trial is the reduction predicted by the model q of f.

slope: current slope

d: potential next direction

xnext: potential next iterate

gnext: current gradient value, if good_grad is true, then this value has been udpated.

robust: if true, try to trap potential cancellation errors

Output:

r: actual vs. predicted reduction radio ∆f/Δq

good_grad

gnext

We assume that q is being minimized, and therefore that Δq > 0.